Update notification. Implement validation

+++ InventoryApp/ViewModels/Product/ProductViewModel.cs
@@ -3,23 +3,49 @@ using System.Collections.ObjectModel;
 using InventoryApp.Models.Product;
 using InventoryApp.Models.Base;
 using System.Linq;
+using InventoryApp.ViewModels.Service;
+using InventoryApp.Models.Service;
 
 namespace InventoryApp.ViewModels.Product
 {
     class ProductViewModel : ViewModelsBase
     {
-        private const string TableName = "Product";
-
-        public ObservableCollection<ProductModel> ProductModels { get; set; }
-        public RelayCommand DeleteCommand { get; set; }
-        public ProductModel SelectedItem { get; set; }
-
         public ProductViewModel()
         {
-            ProductModels = new ObservableCollection<ProductModel>();           
+            ProductModels = new ObservableCollection<ProductModel>();
             DeleteCommand = new RelayCommand((obj) => Delete());
+            AddCommand = new RelayCommand((obj) => Add());
+            AddNewProduct = new ProductModel();
+            Notification = new NotificationServiceViewModel();
+            DataBaseStaticModels = new DataBaseStaticModels();
+            ModelValidation = new ValidationViewModel<ProductModel>();
             Update();
         }
+
+        #region Properties
+        private const string TableName = "Product";
+        public ObservableCollection<ProductModel> ProductModels { get; set; }
+        public DataBaseStaticModels DataBaseStaticModels { get; set; }
+        private ValidationViewModel<ProductModel> ModelValidation { get; set; }
+        public RelayCommand DeleteCommand { get; set; }
+        public RelayCommand AddCommand { get; set; }
+
+        private ProductModel selectedItem;
+        public ProductModel SelectedItem
+        {
+            get => selectedItem;
+            set
+            {
+                if (value != selectedItem)
+                {
+                    selectedItem = value;
+                    OnPropertyChanged(nameof(SelectedItem));
+                }
+            }
+        }
+
+        public ProductModel AddNewProduct { get; set; }
+        public NotificationServiceViewModel Notification { get; set; }
 
         private string searchText;
         public string SearchText
@@ -43,7 +69,9 @@ namespace InventoryApp.ViewModels.Product
                 }
             }
         }
+        #endregion
 
+        #region Functions
         private void Update()
         {
             ProductModels = new BaseQuery().Fill<ProductModel>(($"Get{TableName}"));
@@ -52,8 +80,45 @@ namespace InventoryApp.ViewModels.Product
 
         private void Delete()
         {
-            new BaseQuery().Delete(TableName, SelectedItem.Id);
-            ProductModels.Remove(SelectedItem);
+            if (SelectedItem?.Id != null)
+            {
+                bool isCompleted = new BaseQuery().Delete(TableName, SelectedItem.Id);
+                if (isCompleted)
+                {
+                    ProductModels.Remove(SelectedItem);
+                    Notification.ShowNotification("Info: Product is deleted.");
+                }
+                else
+                {
+                    Notification.ShowNotification("Error: Deleting product failed.");
+                }
+            }
+            else
+            {
+                Notification.ShowNotification("Error: No products selected.");
+            }
+        }
+
+        private void Add()
+        {
+            var errorList = ModelValidation.ValidateFields(AddNewProduct);
+            if (errorList.Any())
+            {
+                Notification.ShowListNotification(errorList);
+            }
+            else
+            {
+                bool isCompleted = new BaseQuery().Add(TableName, AddNewProduct);
+                if (isCompleted)
+                {
+                    ProductModels.Add(AddNewProduct);
+                    Notification.ShowNotification($"Info: {AddNewProduct.Name} is added.");
+                }
+                else
+                {
+                    Notification.ShowNotification("Error: Adding new product failed.");
+                }
+            }
         }
 
         private void Find(string searchText)
@@ -73,5 +138,6 @@ namespace InventoryApp.ViewModels.Product
                 OnPropertyChanged(nameof(ProductModels));
             }
         }
+        #endregion
     }
 }
